Download stock data

Download historical Market Quotations for a list of stock codes or index codes from Net Ease (a web site providing financial information in China, http://money.163.com/).

cntrade(ticker = "", type = "stock")

Arguments

ticker
Ticker is a stock code or index code to be downloaded from Net. In China, stocks are identified by a six digit numbers, not tickers as in the United States. Examples of codes and the names are as following: Stock Codes and Stock Names: 000001 Pingan Bank 000002 Vank Real Estate Co. Ltd. 600000 Pudong Development Bank 600005 Wuhan Steel Co. Ltd. 900901 INESA Electron Co.,Ltd. Index Codes and Index Names: 000001 The Shanghai Composite Index. 000300 CSI 300 Index. 399001 Shenzhen Component Index. The leading zeros in each code can be omitted.

type
Specify that the code is a stock code or index code, 'stock' is the default choice.

Examples

cntrade(1, type = "index")

#> # A tibble: 7,332 x 15
#>    date       code  name  close  high   low  open `yesterday clos…
#>    <date>     <chr> <chr> <dbl> <dbl> <dbl> <dbl>            <dbl>
#>  1 2020-12-16 0000… 上证指数… 3367. 3379. 3359. 3371.            3367.
#>  2 2020-12-15 0000… 上证指数… 3367. 3374. 3348. 3367.            3369.
#>  3 2020-12-14 0000… 上证指数… 3369. 3371. 3339. 3350.            3347.
#>  4 2020-12-11 0000… 上证指数… 3347. 3383. 3325. 3381.            3373.
#>  5 2020-12-10 0000… 上证指数… 3373. 3385. 3358. 3366.            3372.
#>  6 2020-12-09 0000… 上证指数… 3372. 3423. 3372. 3416.            3410.
#>  7 2020-12-08 0000… 上证指数… 3410. 3429. 3403. 3418.            3417.
#>  8 2020-12-07 0000… 上证指数… 3417. 3450. 3414. 3447.            3445.
#>  9 2020-12-04 0000… 上证指数… 3445. 3448. 3417. 3437.            3442.
#> 10 2020-12-03 0000… 上证指数… 3442. 3452. 3429. 3449.            3449.
#> # … with 7,322 more rows, and 7 more variables: `change amount` <dbl>, `change
#> #   percent` <dbl>, turnover <lgl>, transaction <dbl>, volume <dbl>, `total
#> #   market capitalisation` <lgl>, `traded market value` <lgl>